requirements.txt
colorama==0.4.6
PyExifTool==0.5.6
pynput==1.8.0 